tests/systemtests/requirements.txt
PyYAML==5.1.2
pytest-bdd==3.2.1
pytest-xdist==1.30.0
pytest-watch==4.2.0
